start of major file reorganization
