--- core/styles/ewf.less
@@ -1,129 +1,0 @@
-@import "bootstrap-mixins.less";
-
-.edge-web-fonts {
-    .csl {
-        margin: 0px;
-        padding: 0px;
-        font-size: 75%;
-        
-        li {
-            display: inline;
-            line-height: normal;
-        }
-        
-        li:before {
-            content: ", "
-        }
-
-        li:first-child:before {
-            content: ""
-        }
-    }
-    
-    div {
-        overflow: hidden;
-        margin: 0;
-        padding: 0;
-        background-clip: border-box;
-
-    }
-    .ewf-picker {
-        height: 600px;
-        width: 780px;
-        background-color: #f8f8f8;
-        position: relative;
-        
-        .border-radius(5px);
-        .box-shadow(rgba(0, 0, 0, 0.0976563) 1px 2px 5px 0px);
-        border: 1px solid #ccc;
-    }
-    
-    .ewf-tabs {
-        width: 180px;
-        height: 580px;
-        position: absolute;
-        top: 0px;
-        left: 0px;
-        padding: 10px;
-        overflow-y: scroll;
-        background-color: darken(#f8f8f8, 10%);
-        .border-radius(5px 0px 0px 5px);
-        color: #333;
-
-        a:link {
-            color: inherit;
-        }
-        
-        a.selected {
-            text-decoration: underline   
-        }
-        
-        li {
-            margin: 0px;
-            padding: 0px;
-        }
-        
-        ul {
-            margin: 0px;
-            padding: 0px;
-            list-style-type: none;
-        }
-        
-
-    }
-    
-    .ewf-results {
-        position: absolute;
-        left: 200px;
-        width: 580px;
-        max-height:600px;
-        overflow-y: scroll;
-    }
-    
-    .ewf-font {
-        position: relative;
-        width: 120px;
-        height: 120px;
-        top: 0px;
-        left: 0px;
-        float: left;
-        margin-top: 18px;
-        margin-left: 18px;
-
-        .border-radius(5px);
-        .box-shadow(rgba(0, 0, 0, 0.0976563) 1px 2px 5px 0px);
-        border: 1px solid #ccc;
-
-        background-color: #fff;
-        background-repeat: no-repeat;
-        background-position: center 5px;
-        
-        cursor: pointer;
-        
-        &.selected {
-            border-color: #F18900;
-        }
-
-    }
-    
-    .ewf-font-details {
-        position: absolute;
-        top: 80px;
-        left: 0px;
-        width: 114px;
-        height: 40px;
-        padding-left: 3px;
-        padding-right: 3px;
-
-        .border-radius(0px 0px 5px 5px);
-
-        background-color: fade(darken(#f8f8f8, 10%), 75%);
-
-        h1 {
-            font-size: 100%;
-            padding: 0px;
-            margin: 0px;
-            line-height:normal;
-        }
-    }
-}
