+++ ewf-howto-dialog.html
@@ -1,18 +1,18 @@
 <div class="edge-web-fonts-howto-dialog template modal hide">
     <div class="modal-header">
         <a href="#" class="close">&times;</a>
-        <h1 class="dialog-title">{{PRODUCT_NAME}}</h1>
+        <h1 class="dialog-title">{{Strings.PRODUCT_NAME}}</h1>
     </div>
     <div class="modal-body">
-        <p class="dialog-message">{{HOWTO_INSTRUCTIONS_1}}</p>
-        <div class="diagram"><img src="" id="ewf-howto-dialog-image" alt="{{HOWTO_INSTRUCTIONS_1}}"/></div>
+        <p class="dialog-message">{{Strings.HOWTO_INSTRUCTIONS_1}}</p>
+        <div class="ewf-diagram"><img src="{{Paths.ROOT}}img/ewf-howto-dialog.png" alt="{{Strings.HOWTO_INSTRUCTIONS_1}}"/></div>
         <ol>
-            <li>{{HOWTO_INSTRUCTIONS_2}}</li>
-            <li>{{HOWTO_INSTRUCTIONS_3}}</li>
-            <li>{{HOWTO_INSTRUCTIONS_4}}</li>
+            <li>{{Strings.HOWTO_INSTRUCTIONS_2}}</li>
+            <li>{{Strings.HOWTO_INSTRUCTIONS_3}}</li>
+            <li>{{Strings.HOWTO_INSTRUCTIONS_4}}</li>
         </ol>
     </div>
     <div class="modal-footer">
-        <a href="#" class="dialog-button btn primary" data-button-id="ok">{{DIALOG_DONE}}</a>
+        <a href="#" class="dialog-button btn primary" data-button-id="ok">{{Strings.DIALOG_DONE}}</a>
     </div>
 </div>
