+++ img/fonts_decorative@2x.png
