+++ img/popover_arrow@2x.png
