+++ nls/de/strings.js
