+++ nls/fr/strings.js
