+++ nls/root/strings.js
