+++ nls/strings.js
