+++ strings.js
@@ -33,6 +33,6 @@
 define(function (require, exports, module) {
     "use strict";
 
-    module.exports = require("i18n!core/nls/strings");
+    module.exports = require("i18n!nls/strings");
 
 });
